requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==5.2.8
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
python-decouple==3.8
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
